PRSD-958: Creates base ExemptionReasonFormPage
