+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/pages/basePages/ExemptionReasonFormPage.kt
@@ -1,0 +1,23 @@
+package uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages
+
+import com.microsoft.playwright.Page
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.Form
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.Radios
+
+abstract class ExemptionReasonFormPage<ReasonType : Enum<ReasonType>>(
+    page: Page,
+    urlSegment: String,
+) : BasePage(page, urlSegment) {
+    val form = ExemptionReasonForm(page)
+
+    fun submitExemptionReason(exemptionReason: ReasonType) {
+        form.exemptionReasonRadios.selectValue(exemptionReason)
+        form.submit()
+    }
+
+    class ExemptionReasonForm(
+        page: Page,
+    ) : Form(page) {
+        val exemptionReasonRadios = Radios(locator)
+    }
+}
